requirements.txt
Flask==2.3.2
boto3==1.26.142
mysql-connector-python==8.0.33